Implement custom "Natural timeout"
"Natural timeout" is the time duration for which a Hover
notification stays on screen if there is no interaction with
the device or the notification.

Methods that implement "Notification waiting" time delay,
i.e. the duration for which Hover notification stays on screen
if another Hover notification is waiting to be shown, are unused
in Hover.java. Ground work for hooking those methods have
been laid should they be used in future.

Signed-off-by: Shubhang Rathore <shubhang92@gmail.com>