Add "Test notification" feature
After 3 seconds, a test notification is posted to the status bar.

Time delay has been given so that the user may exit from xHover
back to home screen or any other app to be able to send the
notification to Hover. If the user stays in xHover while the test
notification is being posted, the notification doesn't show up in
Hover. That is by design, as made by Paranoid Android developers.

Signed-off-by: Shubhang Rathore <shubhang92@gmail.com>